pyproject.toml
[tool.poetry]
name = "django-template1"
version = "0.1.0"
description = ""
authors = ["pekshevds <pekshev.ds@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.4.0"
django-filter = "^24.3"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
